requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
fastapi==0.85.1
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
python-decouple==3.6
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
